feat: add stylesheet for the html
Maximum height and width of the parent container was specified. Using this,
another style for the children of the parent container was added. This
changes the behavior of the children to consume the remaining space of the
parent container. While at the same time not allowing the children to
overflow.