refactor: change how the function add divs
Previously the function was used to make only 16 divs not following the
the project guidelines. Reworked the code so that 16x16 divs will be
made. This was possible as there will be 16 parent containers and inside
each of them will have 16 children. Using the css, this will become 16x16
grid.